fix storyboard
